+++ tests/test_pandas.py
@@ -289,8 +289,4 @@ def test_datetime_decide_from_contexts_df():
 def test_tree_visualization():
   tree1 = CLIENT.get_decision_tree(AGENT_ID,
                                    DATETIME_AGENT_DATA.last_valid_index().value // 10 ** 9)
-  tree2 = CLIENT.get_decision_tree(AGENT_ID,
-                                   DATETIME_AGENT_DATA.last_valid_index().value // 10 ** 9)
-  html1 = craftai.pandas.utils.create_tree_html(tree1)
-  html2 = craftai.pandas.utils.create_tree_html(tree2)
-  assert_equal(html1, html2)
+  craftai.pandas.utils.create_tree_html(tree1, "", "constant", None, 500)